requirements.txt
appdirs==1.4.3
appnope==0.1.0
awsebcli==3.10.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
bleach==2.0.0
blessed==1.14.1
botocore==1.5.23
cement==2.8.2
certifi==2017.4.17
click==6.7
colorama==0.3.7
configparser==3.5.0
cycler==0.10.0
decorator==4.0.11
Django==1.9.2
django-crispy-forms==1.6.1
djangorestframework==3.6.2
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
dotenv==0.0.5
entrypoints==0.2.2
enum34==1.1.6
functools32==3.2.3.post2
html5lib==0.999999999
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
Jinja2==2.9.6
jmespath==0.9.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
MarkupSafe==1.0
matplotlib==2.0.0
mistune==0.7.4
mpmath==0.19
mysql-connector==2.1.4
MySQL-python==1.2.5
mysqlclient==1.3.10
nbconvert==5.1.1
nbformat==4.3.0
nose==1.3.7
notebook==5.0.0
numpy==1.12.1
packaging==16.8
pandas==0.19.2
pandocfilters==1.4.1
pathlib2==2.2.1
pathspec==0.5.0
pexpect==4.2.1
pickleshare==0.7.4
pip-autoremove==0.9.0
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-dotenv==0.6.4
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.9.1
scandir==1.5
scipy==0.19.0
semantic-version==2.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
subprocess32==3.2.7
sympy==1.0
tabulate==0.7.5
termcolor==1.1.0
terminado==0.6
testpath==0.3
tornado==4.5.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.40.0
widgetsnbextension==2.0.0
